@@ -13,6 +13,15 @@
     "format:check": "prettier --check ."
   },
   "dependencies": {
+    "@assistant-ui/react": "^0.11.51",
+    "@assistant-ui/react-markdown": "^0.11.8",
+    "@assistant-ui/react-syntax-highlighter": "^0.11.8",
+    "@codemirror/lang-cpp": "^6.0.3",
+    "@codemirror/lang-css": "^6.3.1",
+    "@codemirror/lang-html": "^6.4.11",
+    "@codemirror/lang-java": "^6.0.2",
+    "@codemirror/lang-javascript": "^6.2.4",
+    "@codemirror/lang-python": "^6.2.1",
     "@langchain/core": "^0.3.42",
     "@langchain/langgraph": "^0.2.55",
     "@langchain/langgraph-api": "^0.0.16",
@@ -20,17 +29,26 @@
     "@langchain/langgraph-sdk": "^0.0.57",
     "@radix-ui/react-avatar": "^1.1.3",
     "@radix-ui/react-dialog": "^1.1.6",
+    "@radix-ui/react-dropdown-menu": "^2.1.16",
+    "@radix-ui/react-icons": "^1.3.2",
     "@radix-ui/react-label": "^2.1.2",
+    "@radix-ui/react-scroll-area": "^1.2.10",
+    "@radix-ui/react-select": "^2.2.6",
     "@radix-ui/react-separator": "^1.1.2",
     "@radix-ui/react-slot": "^1.1.2",
     "@radix-ui/react-switch": "^1.1.3",
     "@radix-ui/react-tooltip": "^1.1.8",
+    "@tailwindcss/typography": "^0.5.19",
+    "@uiw/react-codemirror": "^4.25.4",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "date-fns": "^4.1.0",
+    "docx": "^9.5.1",
     "esbuild": "^0.25.0",
     "esbuild-plugin-tailwindcss": "^2.0.1",
+    "file-saver": "^2.0.5",
     "framer-motion": "^12.4.9",
+    "jspdf": "^3.0.4",
     "katex": "^0.16.21",
     "lodash": "^4.17.21",
     "lucide-react": "^0.476.0",
@@ -40,9 +58,12 @@
     "prettier": "^3.5.2",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-dropzone": "^14.3.8",
     "react-markdown": "^10.0.1",
+    "react-resizable-panels": "^2.1.9",
     "react-syntax-highlighter": "^15.5.0",
     "recharts": "^2.15.1",
+    "rehype-highlight": "^7.0.2",
     "rehype-katex": "^7.0.1",
     "remark-gfm": "^4.0.1",
     "remark-math": "^6.0.0",
@@ -56,6 +77,7 @@
   "devDependencies": {
     "@eslint/js": "^9.19.0",
     "@tailwindcss/postcss": "^4.0.13",
+    "@types/file-saver": "^2.0.7",
     "@types/lodash": "^4.17.16",
     "@types/node": "^22.13.5",
     "@types/react": "^19.0.8",
@@ -68,7 +90,7 @@
     "eslint-plugin-react-hooks": "^5.0.0",
     "eslint-plugin-react-refresh": "^0.4.18",
     "globals": "^15.14.0",
-    "next": "^15.2.3",
+    "next": "16.0.10",
     "postcss": "^8.5.3",
     "tailwind-scrollbar": "^4.0.1",
     "tailwindcss": "^4.0.13",
