@@ -1,5 +1,5 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
+export default {
   darkMode: ["class"],
   content: [
     "./index.html",
@@ -61,5 +61,8 @@ module.exports = {
       },
     },
   },
-  plugins: [require("tailwindcss-animate"), require("tailwind-scrollbar")],
+  plugins: [
+    (await import("tailwindcss-animate")).default,
+    (await import("tailwind-scrollbar")).default,
+  ],
 };
