@@ -1,215 +1,78 @@
-import { v4 as uuidv4 } from "uuid";
-import { ReactNode, useEffect, useRef } from "react";
-import { motion } from "framer-motion";
-import { cn } from "@/lib/utils";
-import { useStreamContext } from "@/providers/Stream";
-import { useState, FormEvent } from "react";
+"use client";
+
+import { ThreadPrimitive } from "@assistant-ui/react";
+import { useQueryState, parseAsBoolean } from "nuqs";
 import { Button } from "../ui/button";
-import { Checkpoint, Message } from "@langchain/langgraph-sdk";
-import { AssistantMessage, AssistantMessageLoading } from "./messages/ai";
-import { HumanMessage } from "./messages/human";
 import {
-  DO_NOT_RENDER_ID_PREFIX,
-  ensureToolCallsHaveResponses,
-} from "@/lib/ensure-tool-responses";
-import { LangGraphLogoSVG } from "../icons/langgraph";
-import { TooltipIconButton } from "./tooltip-icon-button";
-import {
-  ArrowDown,
-  LoaderCircle,
   PanelRightOpen,
   PanelRightClose,
   SquarePen,
+  PanelLeft,
+  ArrowDown,
 } from "lucide-react";
-import { useQueryState, parseAsBoolean } from "nuqs";
-import { StickToBottom, useStickToBottomContext } from "use-stick-to-bottom";
+import { LangGraphLogoSVG } from "../icons/langgraph";
+import { TooltipIconButton } from "@/components/ui/tooltip-icon-button";
+import { Composer } from "./Composer";
+import { UserMessage, AssistantMessage } from "./MessagePrimitives";
 import ThreadHistory from "./history";
-import { toast } from "sonner";
+import { CanvasLayout } from "../canvas/CanvasLayout";
+import { useCanvas } from "@/contexts/CanvasContext";
+import { CreateFileDialog } from "./CreateFileDialog";
+import { FileUploadDialog } from "./FileUploadDialog";
 import { useMediaQuery } from "@/hooks/useMediaQuery";
-import { Label } from "../ui/label";
-import { Switch } from "../ui/switch";
-import { GitHubSVG } from "../icons/github";
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipProvider,
-  TooltipTrigger,
-} from "../ui/tooltip";
+import { cn } from "@/lib/utils";
+import { useEffect } from "react";
+import { motion } from "framer-motion";
 
-function StickyToBottomContent(props: {
-  content: ReactNode;
-  footer?: ReactNode;
-  className?: string;
-  contentClassName?: string;
-}) {
-  const context = useStickToBottomContext();
+const ThreadScrollToBottom = () => {
   return (
-    <div
-      ref={context.scrollRef}
-      style={{ width: "100%", height: "100%" }}
-      className={props.className}
-    >
-      <div ref={context.contentRef} className={props.contentClassName}>
-        {props.content}
-      </div>
-
-      {props.footer}
-    </div>
+    <ThreadPrimitive.ScrollToBottom asChild>
+      <TooltipIconButton
+        tooltip="Scroll to bottom"
+        variant="outline"
+        className="absolute bottom-20 left-1/2 -translate-x-1/2 rounded-full"
+      >
+        <ArrowDown className="w-4 h-4" />
+      </TooltipIconButton>
+    </ThreadPrimitive.ScrollToBottom>
   );
-}
-
-function ScrollToBottom(props: { className?: string }) {
-  const { isAtBottom, scrollToBottom } = useStickToBottomContext();
-
-  if (isAtBottom) return null;
-  return (
-    <Button
-      variant="outline"
-      className={props.className}
-      onClick={() => scrollToBottom()}
-    >
-      <ArrowDown className="w-4 h-4" />
-      <span>Scroll to bottom</span>
-    </Button>
-  );
-}
-
-function OpenGitHubRepo() {
-  return (
-    <TooltipProvider>
-      <Tooltip>
-        <TooltipTrigger asChild>
-          <a
-            href="https://github.com/langchain-ai/agent-chat-ui"
-            target="_blank"
-            className="flex items-center justify-center"
-          >
-            <GitHubSVG width="24" height="24" />
-          </a>
-        </TooltipTrigger>
-        <TooltipContent side="left">
-          <p>Open GitHub repo</p>
-        </TooltipContent>
-      </Tooltip>
-    </TooltipProvider>
-  );
-}
+};
 
 export function Thread() {
   const [threadId, setThreadId] = useQueryState("threadId");
   const [chatHistoryOpen, setChatHistoryOpen] = useQueryState(
     "chatHistoryOpen",
-    parseAsBoolean.withDefault(false),
+    parseAsBoolean.withDefault(false)
   );
-  const [hideToolCalls, setHideToolCalls] = useQueryState(
-    "hideToolCalls",
-    parseAsBoolean.withDefault(false),
+  const [canvasOpen, setCanvasOpen] = useQueryState(
+    "canvasOpen",
+    parseAsBoolean.withDefault(false)
   );
-  const [input, setInput] = useState("");
-  const [firstTokenReceived, setFirstTokenReceived] = useState(false);
   const isLargeScreen = useMediaQuery("(min-width: 1024px)");
+  const { currentFile, loadThreadFiles } = useCanvas();
 
-  const stream = useStreamContext();
-  const { assistantId, setAssistantId } = stream;
-  const messages = stream.messages;
-  const isLoading = stream.isLoading;
-
-  const lastError = useRef<string | undefined>(undefined);
-
+  // Load thread files
   useEffect(() => {
-    if (!stream.error) {
-      lastError.current = undefined;
-      return;
+    if (threadId) {
+      loadThreadFiles(threadId);
     }
-    try {
-      const message = (stream.error as any).message;
-      if (!message || lastError.current === message) {
-        // Message has already been logged. do not modify ref, return early.
-        return;
-      }
+  }, [threadId, loadThreadFiles]);
 
-      // Message is defined, and it has not been logged yet. Save it, and send the error
-      lastError.current = message;
-      toast.error("An error occurred. Please try again.", {
-        description: (
-          <p>
-            <strong>Error:</strong> <code>{message}</code>
-          </p>
-        ),
-        richColors: true,
-        closeButton: true,
-      });
-    } catch {
-      // no-op
-    }
-  }, [stream.error]);
-
-  // TODO: this should be part of the useStream hook
-  const prevMessageLength = useRef(0);
+  // Auto-open canvas when file is created
   useEffect(() => {
-    if (
-      messages.length !== prevMessageLength.current &&
-      messages?.length &&
-      messages[messages.length - 1].type === "ai"
-    ) {
-      setFirstTokenReceived(true);
+    if (currentFile) {
+      setCanvasOpen(true);
     }
+  }, [currentFile, setCanvasOpen]);
 
-    prevMessageLength.current = messages.length;
-  }, [messages]);
-
-  const handleSubmit = (e: FormEvent) => {
-    e.preventDefault();
-    if (!input.trim() || isLoading) return;
-    setFirstTokenReceived(false);
-
-    const newHumanMessage: Message = {
-      id: uuidv4(),
-      type: "human",
-      content: input,
-    };
-
-    const toolMessages = ensureToolCallsHaveResponses(stream.messages);
-    stream.submit(
-      { messages: [...toolMessages, newHumanMessage] },
-      {
-        streamMode: ["values"],
-        optimisticValues: (prev) => ({
-          ...prev,
-          messages: [
-            ...(prev.messages ?? []),
-            ...toolMessages,
-            newHumanMessage,
-          ],
-        }),
-      },
-    );
-
-    setInput("");
-  };
-
-  const handleRegenerate = (
-    parentCheckpoint: Checkpoint | null | undefined,
-  ) => {
-    // Do this so the loading state is correct
-    prevMessageLength.current = prevMessageLength.current - 1;
-    setFirstTokenReceived(false);
-    stream.submit(undefined, {
-      checkpoint: parentCheckpoint,
-      streamMode: ["values"],
-    });
-  };
-
-  const chatStarted = !!threadId || !!messages.length;
-  const hasNoAIOrToolMessages = !messages.find(
-    (m) => m.type === "ai" || m.type === "tool",
-  );
+  const chatStarted = !!threadId;
 
   return (
     <div className="flex w-full h-screen overflow-hidden">
+      {/* Chat History Sidebar */}
       <div className="relative lg:flex hidden">
         <motion.div
-          className="absolute h-full border-r bg-white overflow-hidden z-20"
+          className="absolute h-full border-r border-border bg-background overflow-hidden"
           style={{ width: 300 }}
           animate={
             isLargeScreen
@@ -228,32 +91,21 @@ export function Thread() {
           </div>
         </motion.div>
       </div>
-      <motion.div
+
+      {/* Main Content */}
+      <div
         className={cn(
-          "flex-1 flex flex-col min-w-0 overflow-hidden relative",
-          !chatStarted && "grid-rows-[1fr]",
+          "flex-1 flex flex-col min-w-0 overflow-hidden",
+          chatHistoryOpen && isLargeScreen && "ml-[300px]"
         )}
-        layout={isLargeScreen}
-        animate={{
-          marginLeft: chatHistoryOpen ? (isLargeScreen ? 300 : 0) : 0,
-          width: chatHistoryOpen
-            ? isLargeScreen
-              ? "calc(100% - 300px)"
-              : "100%"
-            : "100%",
-        }}
-        transition={
-          isLargeScreen
-            ? { type: "spring", stiffness: 300, damping: 30 }
-            : { duration: 0 }
-        }
       >
+        {/* Header - Before Chat Started */}
         {!chatStarted && (
-          <div className="absolute top-0 left-0 w-full flex items-center justify-between gap-3 p-2 pl-4 z-10">
-            <div>
+          <div className="absolute top-0 left-0 w-full flex items-center justify-between gap-3 p-2 pl-4 z-30 bg-background border-b border-border">
+            <div className="flex items-center gap-2">
               {(!chatHistoryOpen || !isLargeScreen) && (
                 <Button
-                  className="hover:bg-gray-100"
+                  className="hover:bg-secondary"
                   variant="ghost"
                   onClick={() => setChatHistoryOpen((p) => !p)}
                 >
@@ -265,55 +117,61 @@ export function Thread() {
                 </Button>
               )}
             </div>
-            <div className="absolute top-2 right-4 flex items-center">
-              <OpenGitHubRepo />
+            <div className="absolute top-2 right-4 flex items-center gap-2">
+              <TooltipIconButton
+                tooltip="Open Canvas"
+                variant="ghost"
+                onClick={() => setCanvasOpen(true)}
+              >
+                <PanelLeft className="size-5" />
+              </TooltipIconButton>
             </div>
           </div>
         )}
+
+        {/* Header - After Chat Started */}
         {chatStarted && (
-          <div className="flex items-center justify-between gap-3 p-2 z-10 relative">
-            <div className="flex items-center justify-start gap-2 relative">
-              <div className="absolute left-0 z-10">
-                {(!chatHistoryOpen || !isLargeScreen) && (
-                  <Button
-                    className="hover:bg-gray-100"
-                    variant="ghost"
-                    onClick={() => setChatHistoryOpen((p) => !p)}
-                  >
-                    {chatHistoryOpen ? (
-                      <PanelRightOpen className="size-5" />
-                    ) : (
-                      <PanelRightClose className="size-5" />
-                    )}
-                  </Button>
-                )}
-              </div>
-              <motion.button
+          <div className="flex items-center justify-between gap-3 p-3 border-b border-border z-30 bg-background flex-shrink-0">
+            <div className="flex items-center justify-start gap-2">
+              {(!chatHistoryOpen || !isLargeScreen) && (
+                <Button
+                  className="hover:bg-secondary"
+                  variant="ghost"
+                  onClick={() => setChatHistoryOpen((p) => !p)}
+                >
+                  {chatHistoryOpen ? (
+                    <PanelRightOpen className="size-5" />
+                  ) : (
+                    <PanelRightClose className="size-5" />
+                  )}
+                </Button>
+              )}
+              <button
                 className="flex gap-2 items-center cursor-pointer"
                 onClick={() => setThreadId(null)}
-                animate={{
-                  marginLeft: !chatHistoryOpen ? 48 : 0,
-                }}
-                transition={{
-                  type: "spring",
-                  stiffness: 300,
-                  damping: 30,
-                }}
               >
                 <LangGraphLogoSVG width={32} height={32} />
                 <span className="text-xl font-semibold tracking-tight">
                   Agent Chat
                 </span>
-              </motion.button>
+              </button>
+              {threadId && (
+                <div className="flex gap-2 ml-2">
+                  <CreateFileDialog threadId={threadId} />
+                  <FileUploadDialog threadId={threadId} />
+                </div>
+              )}
             </div>
 
             <div className="flex items-center gap-4">
-              <div className="flex items-center">
-                <OpenGitHubRepo />
-              </div>
               <TooltipIconButton
-                size="lg"
-                className="p-4"
+                tooltip="Open Canvas"
+                variant="ghost"
+                onClick={() => setCanvasOpen(true)}
+              >
+                <PanelLeft className="size-5" />
+              </TooltipIconButton>
+              <TooltipIconButton
                 tooltip="New thread"
                 variant="ghost"
                 onClick={() => setThreadId(null)}
@@ -321,149 +179,88 @@ export function Thread() {
                 <SquarePen className="size-5" />
               </TooltipIconButton>
             </div>
-
-            <div className="absolute inset-x-0 top-full h-5 bg-gradient-to-b from-background to-background/0" />
           </div>
         )}
 
-        <StickToBottom className="relative flex-1 overflow-hidden">
-          <StickyToBottomContent
-            className={cn(
-              "absolute px-4 inset-0 overflow-y-scroll [&::-webkit-scrollbar]:w-1.5 [&::-webkit-scrollbar-thumb]:rounded-full [&::-webkit-scrollbar-thumb]:bg-gray-300 [&::-webkit-scrollbar-track]:bg-transparent",
-              !chatStarted && "flex flex-col items-stretch mt-[25vh]",
-              chatStarted && "grid grid-rows-[1fr_auto]",
-            )}
-            contentClassName="pt-8 pb-16  max-w-3xl mx-auto flex flex-col gap-4 w-full"
-            content={
-              <>
-                {messages
-                  .filter((m) => !m.id?.startsWith(DO_NOT_RENDER_ID_PREFIX))
-                  .map((message, index) =>
-                    message.type === "human" ? (
-                      <HumanMessage
-                        key={message.id || `${message.type}-${index}`}
-                        message={message}
-                        isLoading={isLoading}
-                      />
-                    ) : (
-                      <AssistantMessage
-                        key={message.id || `${message.type}-${index}`}
-                        message={message}
-                        isLoading={isLoading}
-                        handleRegenerate={handleRegenerate}
-                      />
-                    ),
-                  )}
-                {/* Special rendering case where there are no AI/tool messages, but there is an interrupt.
-                    We need to render it outside of the messages list, since there are no messages to render */}
-                {hasNoAIOrToolMessages && !!stream.interrupt && (
-                  <AssistantMessage
-                    key="interrupt-msg"
-                    message={undefined}
-                    isLoading={isLoading}
-                    handleRegenerate={handleRegenerate}
-                  />
+        {/* Chat Thread or Canvas */}
+        {!canvasOpen && (
+          <ThreadPrimitive.Root className="flex-1 flex flex-col overflow-hidden">
+            <ThreadPrimitive.Viewport className="flex-1 overflow-y-auto px-4">
+              <div
+                className={cn(
+                  "max-w-3xl mx-auto flex flex-col gap-4 w-full",
+                  !chatStarted ? "pt-[25vh] pb-8" : "py-8"
                 )}
-                {isLoading && !firstTokenReceived && (
-                  <AssistantMessageLoading />
-                )}
-              </>
-            }
-            footer={
-              <div className="sticky flex flex-col items-center gap-8 bottom-0 bg-white">
+              >
                 {!chatStarted && (
-                  <div className="flex gap-3 items-center">
-                    <LangGraphLogoSVG className="flex-shrink-0 h-8" />
-                    <h1 className="text-2xl font-semibold tracking-tight">
+                  <div className="flex flex-col items-center gap-4 mb-8">
+                    <LangGraphLogoSVG className="h-12" />
+                    <h1 className="text-3xl font-semibold tracking-tight text-foreground">
                       Agent Chat
                     </h1>
+                    <p className="text-muted-foreground text-center">
+                      Start a conversation with the AI assistant
+                    </p>
                   </div>
                 )}
 
-                <ScrollToBottom className="absolute bottom-full left-1/2 -translate-x-1/2 mb-4 animate-in fade-in-0 zoom-in-95" />
-
-                <div className="bg-muted rounded-2xl border shadow-xs mx-auto mb-8 w-full max-w-3xl relative z-10">
-                  <form
-                    onSubmit={handleSubmit}
-                    className="grid grid-rows-[1fr_auto] gap-2 max-w-3xl mx-auto"
-                  >
-                    <textarea
-                      value={input}
-                      onChange={(e) => setInput(e.target.value)}
-                      onKeyDown={(e) => {
-                        if (
-                          e.key === "Enter" &&
-                          !e.shiftKey &&
-                          !e.metaKey &&
-                          !e.nativeEvent.isComposing
-                        ) {
-                          e.preventDefault();
-                          const el = e.target as HTMLElement | undefined;
-                          const form = el?.closest("form");
-                          form?.requestSubmit();
-                        }
-                      }}
-                      placeholder="Type your message..."
-                      className="p-3.5 pb-0 border-none bg-transparent field-sizing-content shadow-none ring-0 outline-none focus:outline-none focus:ring-0 resize-none"
-                    />
-
-                    <div className="flex items-center justify-between p-2 pt-4">
-                      <div className="flex items-center gap-4">
-                        <div className="flex items-center space-x-2">
-                          <Label
-                            htmlFor="assistant-select"
-                            className="text-sm text-gray-600 font-medium"
-                          >
-                            Assistant:
-                          </Label>
-                          <select
-                            id="assistant-select"
-                            value={assistantId || "agent"}
-                            onChange={(e) => setAssistantId(e.target.value)}
-                            className="text-sm border border-gray-200 rounded-lg px-3 py-1.5 bg-white text-gray-800 hover:border-gray-300 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all cursor-pointer shadow-sm"
-                          >
-                            <option value="agent">🤖 Agent</option>
-                            <option value="assistant">💼 Assistant</option>
-                            <option value="chatbot">💬 Chatbot</option>
-                            <option value="helper">🛠️ Helper</option>
-                          </select>
-                        </div>
-                        <div className="flex items-center space-x-2">
-                          <Switch
-                            id="render-tool-calls"
-                            checked={hideToolCalls ?? false}
-                            onCheckedChange={setHideToolCalls}
-                          />
-                          <Label
-                            htmlFor="render-tool-calls"
-                            className="text-sm text-gray-600"
-                          >
-                            Hide Tool Calls
-                          </Label>
-                        </div>
-                      </div>
-                      {stream.isLoading ? (
-                        <Button key="stop" onClick={() => stream.stop()}>
-                          <LoaderCircle className="w-4 h-4 animate-spin" />
-                          Cancel
-                        </Button>
-                      ) : (
-                        <Button
-                          type="submit"
-                          className="transition-all shadow-md"
-                          disabled={isLoading || !input.trim()}
-                        >
-                          Send
-                        </Button>
-                      )}
-                    </div>
-                  </form>
-                </div>
+                <ThreadPrimitive.Messages
+                  components={{
+                    UserMessage: UserMessage,
+                    AssistantMessage: AssistantMessage,
+                  }}
+                />
               </div>
+            </ThreadPrimitive.Viewport>
+
+            <div className="border-t border-border bg-background p-4">
+              <div className="max-w-3xl mx-auto">
+                <Composer />
+              </div>
+            </div>
+
+            <ThreadScrollToBottom />
+          </ThreadPrimitive.Root>
+        )}
+
+        {canvasOpen && threadId && (
+          <CanvasLayout
+            chatPanel={
+              <ThreadPrimitive.Root className="h-full flex flex-col">
+                <ThreadPrimitive.Viewport className="flex-1 overflow-y-auto px-4 py-8">
+                  <div className="max-w-2xl mx-auto flex flex-col gap-4">
+                    <ThreadPrimitive.Messages
+                      components={{
+                        UserMessage: UserMessage,
+                        AssistantMessage: AssistantMessage,
+                      }}
+                    />
+                  </div>
+                </ThreadPrimitive.Viewport>
+                <div className="border-t border-border bg-background p-4">
+                  <div className="max-w-2xl mx-auto">
+                    <Composer />
+                  </div>
+                </div>
+              </ThreadPrimitive.Root>
             }
+            threadId={threadId}
           />
-        </StickToBottom>
-      </motion.div>
+        )}
+      </div>
+
+      {/* Toggle Canvas Button */}
+      {!canvasOpen && currentFile && (
+        <div className="fixed bottom-20 right-6 z-50">
+          <TooltipIconButton
+            tooltip="Open Canvas"
+            onClick={() => setCanvasOpen(true)}
+            className="shadow-xl bg-primary text-primary-foreground hover:bg-primary/90 size-12"
+          >
+            <PanelLeft className="size-6" />
+          </TooltipIconButton>
+        </div>
+      )}
     </div>
   );
 }
