@@ -19,7 +19,7 @@ const buttonVariants = cva(
           "bg-secondary text-secondary-foreground shadow-xs hover:bg-secondary/80",
         ghost: "hover:bg-accent hover:text-accent-foreground",
         link: "text-primary underline-offset-4 hover:underline",
-        brand: "bg-[#2F6868] hover:bg-[#2F6868]/90 border-[#2F6868] text-white",
+        brand: "bg-brand hover:bg-brand/90 border-brand text-brand-foreground",
       },
       size: {
         default: "h-9 px-4 py-2 has-[>svg]:px-3",
